feat: updated darknode gas component
