--- src/controllers/operatorPages/darknodePage/blocks/GasBlock.tsx
@@ -1,87 +1,0 @@
-import BigNumber from "bignumber.js";
-import React, { useCallback, useState } from "react";
-
-import {
-    DarknodesState,
-    NetworkContainer,
-} from "../../../../store/networkContainer";
-import { ReactComponent as FlameIcon } from "../../../../styles/images/icon-flame.svg";
-import { TopUpController } from "../topup/TopUpController";
-import { Block, BlockTitle } from "./Block";
-
-interface Props {
-    darknodeDetails: DarknodesState | null;
-}
-
-export const GasBlock: React.FC<Props> = ({ darknodeDetails }) => {
-    const gasValue =
-        darknodeDetails && darknodeDetails.ethBalance
-            ? darknodeDetails.ethBalance
-                  .div(new BigNumber(Math.pow(10, 18)))
-                  .toFixed(3)
-            : null;
-
-    const { updateDarknodeDetails } = NetworkContainer.useContainer();
-
-    const [loading, setLoading] = useState(false);
-
-    const darknodeID = darknodeDetails && darknodeDetails.ID;
-
-    const reloadDetails = useCallback(async (): Promise<void> => {
-        setLoading(true);
-        try {
-            if (darknodeID) {
-                await updateDarknodeDetails(darknodeID);
-            }
-        } catch (error) {
-            console.error(error);
-        }
-        setLoading(false);
-    }, [updateDarknodeDetails, darknodeID]);
-
-    return (
-        <Block
-            // className={`gas-block ${showAdvanced ? "" : "basic"}`}
-            className={`gas-block`}
-            // onClick={showAdvanced ? undefined : this.toggleAdvanced}
-        >
-            <BlockTitle>
-                <h3>
-                    <FlameIcon />
-                    Gas Balance
-                </h3>
-            </BlockTitle>
-
-            {darknodeDetails ? (
-                <div className="block--advanced">
-                    <div className="block--advanced--top">
-                        {gasValue ? (
-                            <>
-                                <span className="gas-block--advanced--value">
-                                    {gasValue}
-                                </span>
-                                <span className="gas-block--advanced--unit">
-                                    ETH
-                                </span>
-                            </>
-                        ) : (
-                            <>
-                                Unable to load.{" "}
-                                <button
-                                    className="text-button"
-                                    disabled={loading}
-                                    onClick={reloadDetails}
-                                >
-                                    Reload
-                                </button>
-                            </>
-                        )}
-                    </div>
-                    <div className="block--advanced--bottom">
-                        <TopUpController darknodeID={darknodeDetails.ID} />
-                    </div>
-                </div>
-            ) : null}
-        </Block>
-    );
-};
