+++ src/store/popupContainer.ts
@@ -1,5 +1,5 @@
 // import { useState } from "react";
-import { useState } from "react";
+import { useCallback, useState } from "react";
 import { createContainer } from "unstated-next";
 
 interface PopupDetails {
@@ -17,24 +17,27 @@ const usePopupContainer = () => {
     const [dismissible, setDismissible] = useState<boolean>(true);
     const [onCancel, setOnCancel] = useState<() => void>(() => null);
 
-    const setPopup: (details: PopupDetails) => void = ({
-        popup: newPopup,
-        overlay: newOverlay,
-        dismissible: newDismissible,
-        onCancel: newOnCancel,
-    }: PopupDetails) => {
-        setPopupElement(newPopup);
-        setOverlay(newOverlay || false);
-        setDismissible(newDismissible || false);
-        setOnCancel(() => newOnCancel);
-    };
+    const setPopup: (details: PopupDetails) => void = useCallback(
+        ({
+            popup: newPopup,
+            overlay: newOverlay,
+            dismissible: newDismissible,
+            onCancel: newOnCancel,
+        }: PopupDetails) => {
+            setPopupElement(newPopup);
+            setOverlay(newOverlay || false);
+            setDismissible(newDismissible || false);
+            setOnCancel(() => newOnCancel);
+        },
+        [],
+    );
 
-    const clearPopup = () => {
+    const clearPopup = useCallback(() => {
         setPopupElement(null);
         setOverlay(false);
         setDismissible(true);
         setOnCancel(() => null);
-    };
+    }, []);
 
     return {
         setPopup,
